backend/requirements.txt
aiofiles==22.1.0
aiosqlite==0.17.0
anyio==3.6.2
asgiref==3.5.2
asyncpg==0.27.0
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
databases==0.6.2
distlib==0.3.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.0.1
fastapi==0.86.0
fastapi_keycloak==1.0.7
filelock==3.8.0
greenlet==2.0.1
h11==0.14.0
identify==2.5.8
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
natsort==8.2.0
nodeenv==1.7.0
packaging==21.3
passlib==1.7.4
platformdirs==2.5.3
pluggy==1.0.0
pre-commit==2.20.0
psycopg2==2.9.5
pyasn1==0.4.8
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
python-certifi-win32==1.6.1
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
seedir==0.4.0
setuptools-scm==7.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
virtualenv==20.16.7
wrapt==1.14.1